🏗️ Move player state types to a shared module
Moves the player state types out of the main process player state store
into src/shared/player-state.ts so they can be consumed from outside the
main process, and re-exports them from the store to keep every existing
import working. Fields that are genuinely absent before the first track
loads (videoDetails, playlistId, queue, PlayerQueueItem.counterparts) are
now typed as nullable to match the real initial state.

No behaviour change.
